+++ openbank-account-service/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.27.4](https://github.com/JiRaska/open-bank-oss/compare/account-service-v0.27.3...account-service-v0.27.4) (2026-09-03)
+
+
+### Bug Fixes
+
+* **account:** entity createdAt/updatedAt defaults EPOCH → Instant.now() ([#8403](https://github.com/JiRaska/open-bank-oss/issues/8403)) ([c1a3953](https://github.com/JiRaska/open-bank-oss/commit/c1a3953bf54da88540b86f30418fedfdd0509a80))
+
+## [0.27.3](https://github.com/JiRaska/open-bank-oss/compare/account-service-v0.27.2...account-service-v0.27.3) (2026-09-03)
+
+
+### Bug Fixes
+
+* **account:** revoking an unknown authorization answers 404, not 500 ([#8466](https://github.com/JiRaska/open-bank-oss/issues/8466)) ([909dee5](https://github.com/JiRaska/open-bank-oss/commit/909dee547a0bf602792adf60b9df91c28abe55be))
+
 ## [0.27.2](https://github.com/JiRaska/open-bank-oss/compare/account-service-v0.27.1...account-service-v0.27.2) (2026-09-03)